requirements.txt
Flask==3.1.0
Flask-WTF==1.1.1
Werkzeug==3.1.3
bcrypt==4.0.1
SQLAlchemy==2.0.38
ulid-py==1.1.0
PyJWT==2.10.1
Flask-Mail==0.10.0
flask-admin==1.6.1
pytz==2025.1
email_validator==2.2.0
pytest==8.3.5
unittest2==1.1.0